+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## master
 
+## 1.2.2RC2
+* Add Inside Axelerant link
+* Store branding updates
+* Update about Axelerant verbiage
+
 ## 1.2.1
 * RESOLVE PHP Fatal error: Access to undeclared static property: CLASS::$scripts_called